created an annotation function, tests pass
next I just need to create the annotation prompts for the other
categories, and then the next phase of processing will be hard-coded
stuff to do, depending on which category it is. For instance, I will
have a note type called “command” that can do stuff like change the
category on a miscategorized note, which would trigger the loop again on
that note that changed.